+++ DSCResources/MSFT_WaitForVolume/MSFT_WaitForVolume.psm1
@@ -1,24 +1,12 @@
-# Suppressed as per PSSA Rule Severity guidelines for unit/integration tests:
-# https://github.com/PowerShell/DscResources/blob/master/PSSARuleSeverities.md
-[System.Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSUseShouldProcessForStateChangingFunctions', '')]
-param ()
-
 $modulePath = Join-Path -Path (Split-Path -Path (Split-Path -Path $PSScriptRoot -Parent) -Parent) -ChildPath 'Modules'
 
-# Import the Storage Common Modules
+# Import the Storage Common Module.
 Import-Module -Name (Join-Path -Path $modulePath `
-                               -ChildPath (Join-Path -Path 'StorageDsc.Common' `
-                                                     -ChildPath 'StorageDsc.Common.psm1'))
+        -ChildPath (Join-Path -Path 'StorageDsc.Common' `
+            -ChildPath 'StorageDsc.Common.psm1'))
 
-# Import the Storage Resource Helper Module
-Import-Module -Name (Join-Path -Path $modulePath `
-                               -ChildPath (Join-Path -Path 'StorageDsc.ResourceHelper' `
-                                                     -ChildPath 'StorageDsc.ResourceHelper.psm1'))
-
-# Import Localization Strings
-$localizedData = Get-LocalizedData `
-    -ResourceName 'MSFT_WaitForVolume' `
-    -ResourcePath (Split-Path -Parent $Script:MyInvocation.MyCommand.Path)
+# Import Localization Strings.
+$script:localizedData = Get-LocalizedData -ResourceName 'MSFT_WaitForVolume'
 
 <#
     .SYNOPSIS
@@ -54,7 +42,7 @@ function Get-TargetResource
 
     Write-Verbose -Message ( @(
             "$($MyInvocation.MyCommand): "
-            $($localizedData.GettingWaitForVolumeStatusMessage -f $DriveLetter)
+            $($script:localizedData.GettingWaitForVolumeStatusMessage -f $DriveLetter)
         ) -join '' )
 
     # Validate the DriveLetter parameter
@@ -101,7 +89,7 @@ function Set-TargetResource
 
     Write-Verbose -Message ( @(
             "$($MyInvocation.MyCommand): "
-            $($localizedData.CheckingForVolumeStatusMessage -f $DriveLetter)
+            $($script:localizedData.CheckingForVolumeStatusMessage -f $DriveLetter)
         ) -join '' )
 
     # Validate the DriveLetter parameter
@@ -116,7 +104,7 @@ function Set-TargetResource
         {
             Write-Verbose -Message ( @(
                     "$($MyInvocation.MyCommand): "
-                    $($localizedData.VolumeFoundMessage -f $DriveLetter)
+                    $($script:localizedData.VolumeFoundMessage -f $DriveLetter)
                 ) -join '' )
 
             $volumeFound = $true
@@ -126,7 +114,7 @@ function Set-TargetResource
         {
             Write-Verbose -Message ( @(
                     "$($MyInvocation.MyCommand): "
-                    $($localizedData.VolumeNotFoundMessage -f $DriveLetter,$RetryIntervalSec)
+                    $($script:localizedData.VolumeNotFoundRetryingMessage -f $DriveLetter,$RetryIntervalSec)
                 ) -join '' )
 
             Start-Sleep -Seconds $RetryIntervalSec
@@ -140,7 +128,7 @@ function Set-TargetResource
     if (-not $volumeFound)
     {
         New-InvalidOperationException `
-            -Message $($localizedData.VolumeNotFoundAfterError -f $DriveLetter,$RetryCount)
+            -Message $($script:localizedData.VolumeNotFoundAfterError -f $DriveLetter,$RetryCount)
     } # if
 } # function Set-TargetResource
 
@@ -178,22 +166,25 @@ function Test-TargetResource
 
     Write-Verbose -Message ( @(
             "$($MyInvocation.MyCommand): "
-            $($localizedData.TestingWaitForVolumeStatusMessage -f $DriveLetter)
+            $($script:localizedData.CheckingForVolumeStatusMessage -f $DriveLetter)
         ) -join '' )
 
     # Validate the DriveLetter parameter
     $DriveLetter = Assert-DriveLetterValid -DriveLetter $DriveLetter
 
-    # This command forces a refresh of the PS Drive subsystem.
-    # So triggers any "missing" drives to show up.
+    <#
+        This command forces a refresh of the PS Drive subsystem.
+        So triggers any "missing" drives to show up.
+    #>
     $null = Get-PSDrive
 
     $volume = Get-Volume -DriveLetter $DriveLetter -ErrorAction SilentlyContinue
+
     if ($volume)
     {
         Write-Verbose -Message ( @(
                 "$($MyInvocation.MyCommand): "
-                $($localizedData.VolumeFoundMessage -f $DriveLetter)
+                $($script:localizedData.VolumeFoundMessage -f $DriveLetter)
             ) -join '' )
 
         return $true
@@ -201,7 +192,7 @@ function Test-TargetResource
 
     Write-Verbose -Message ( @(
             "$($MyInvocation.MyCommand): "
-            $($localizedData.VolumeNotFoundMessage -f $DriveLetter)
+            $($script:localizedData.VolumeNotFoundMessage -f $DriveLetter)
         ) -join '' )
 
     return $false