+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+- Removed suppression of `PSUseShouldProcessForStateChangingFunctions` PSSA rule
+  because it is no longer required.
+- Combined all `StorageDsc.ResourceHelper` module functions into
+  `StorageDsc.Common` module and removed `StorageDsc.ResourceHelper`.
+- Opted into Common Tests 'Common Tests - Validate Localization' -
+  fixes [Issue #206](https://github.com/PowerShell/StorageDsc/issues/206).
+- Refactored tests for `StorageDsc.Common` to meet latest standards.
+- Minor style corrections.
+- Removed unused localization strings from resources.
+
 ## 4.7.0.0
 
 - DiskAccessPath: